WebApi：项目接口，平台接口，完成。资源包接口未完成

+++ AssetManager.API/Service/IService/IBaseService.cs
@@ -7,12 +7,12 @@ namespace AssetManager.API.Service.IService
     {
         Task<ApiResponse> GetAllAsync(QueryParameter parameter);
 
-        Task<ApiResponse> GetSingleAsync(int id);
+        Task<ApiResponse> GetSingleAsync(long id);
 
         Task<ApiResponse> AddAsync(T model);
 
         Task<ApiResponse> UpdateAsync(T model);
 
-        Task<ApiResponse> DeleteAsync(int id);
+        Task<ApiResponse> DeleteAsync(long id);
     }
 }
